Abstract validation and added to publish call Changes to be committed: 	modified:   api/model/prefix.py 	modified:   api/scripts/method_specific/POST_api_objects_drafts_create.py 	modified:   api/scripts/method_specific/POST_api_objects_publish.py 	modified:   api/scripts/method_specific/POST_validate_payload_against_schema.py 	modified:   api/scripts/utilities/JsonUtils.py 	modified:   api/views.py
