+++ bco_api/api/scripts/utilities/JsonUtils.py
@@ -1,38 +1,168 @@
-# For JSON parsing and schema validation.
+#!/usr/bin/env python3
+"""JSON Utils
+
+For JSON parsing and schema validation.
+"""
+
+import os
+import sys
 import json
 import jsonref
 import jsonschema
+from simplejson.errors import JSONDecodeError
+from requests.exceptions import ConnectionError as ErrorConnecting
 
-# For catching print output.
-import io
-import sys
+
+def get_schema(schema_uri):
+    """Retrieve JSON Schema
+
+    Parameters
+    ----------
+    schema_uri : str
+        A URI that is used to pull the JSON schema for validation.
+
+    Returns
+    -------
+    schema : dict
+        A dictionary of the JSON schema definition, or detail on the error loading the schema.
+    """
+
+    try:
+        schema = jsonref.load_uri(schema_uri)
+        return schema
+
+    except JSONDecodeError:
+        return {schema_uri: ["Failed to load extension schema. JSON Decode Error."]}
+
+    except TypeError:
+        return {schema_uri: ["Failed to load extension schema. Invalid format."]}
+
+    except ErrorConnecting:
+        return {schema_uri: ["Failed to load extension schema. Connection Error."]}
+
+
+def validate(schema, json_object, results):
+    """BCO/extension Validator
+
+    Parameters
+    ----------
+    schema : dict
+        A dictionary of the JSON schema definition.
+    json_object : dict
+        A dictionary of the BCO/extension JSON for validation.
+    results : dict
+        A dictionary that is used to collect the validation results.
+
+    Returns
+    -------
+    results : dict
+       A dictionary that is used to collect the validation results.
+    """
+
+    if "object_id" in json_object:
+        identifier = json_object["object_id"]
+
+    if "extension_schema" in json_object:
+        identifier = json_object["extension_schema"]
+
+    validator = jsonschema.Draft7Validator(schema)
+    errors = validator.iter_errors(json_object)
+    for error in errors:
+        values = "".join(f"[{v}]" for v in error.path)
+        results[identifier]["number_of_errors"] += 1
+        if len(values) == 0:
+            error_string = {"top_level": error.message}
+        else:
+            error_string = {values: error.message}
+        results[identifier]["error_detail"].append(error_string)
+
+    return results
+
+
+def parse_bco(bco, results):
+    """BCO Parsing for Validation
+
+    Parameters
+    ----------
+    bco : JSON
+        The BCO JSON to be processed for validation.
+    results : dict
+        A dictionary to be populated with the BCO validation results
+
+    Returns
+    -------
+    results : dict
+        A dictionary with the BCO validation results
+    """
+
+    identifier = bco["object_id"]
+    results[identifier] = {"number_of_errors": 0, "error_detail": []}
+    try:
+        spec_version = get_schema(bco["spec_version"])
+    except ErrorConnecting:
+        file_path = os.path.dirname(
+            os.path.abspath("api/validation_definitions/IEEE/2791object.json")
+        )
+
+        ieee = "api/validation_definitions/IEEE/2791object.json"
+        with open(ieee, "r", encoding="utf-8") as file:
+            spec_version = jsonref.load(
+                file, base_uri=f"file://{file_path}/", jsonschema=True
+            )
+    results = validate(spec_version, bco, results)
+    if "extension_domain" in bco.keys():
+        for extension in bco["extension_domain"]:
+            extension_id = extension["extension_schema"]
+            results[identifier][extension_id] = {
+                "number_of_errors": 0,
+                "error_detail": [],
+            }
+            extension_schema = get_schema(extension_id)
+            if extension_id in extension_schema:
+                results[identifier][extension_id] = {
+                    "number_of_errors": 1,
+                    "error_detail": extension_schema,
+                }
+            else:
+                results[identifier] = validate(
+                    extension_schema, extension, results[identifier]
+                )
+            if results[identifier][extension_id]["number_of_errors"] == 0:
+                results[identifier][extension_id]["error_detail"] = ["Extension Valid"]
+
+            results[identifier]["number_of_errors"] += results[identifier][
+                extension_id
+            ]["number_of_errors"]
+
+    return results
 
 
 class JsonUtils:
+    """Class Description
+    -----------------
 
-    # Class Description
-    # -----------------
-
-    # These are methods for checking for valid JSON objects.
+    These are methods for checking for valid JSON objects.
+    """
 
     # Check for a set of keys.
     def check_key_set_exists(self, data_pass, key_set):
+        """
+        Arguments
+        ---------
 
-        # Arguments
-        # ---------
+        data_pass: the 'raw' data.
 
-        # data_pass: the 'raw' data.
+        Go over each key in the key set and see if it exists
+        the in request data.
 
-        # Go over each key in the key set and see if it exists
-        # the in request data.
+        Returns
+        -------
 
-        # Returns
-        # -------
+        None: all keys were present
+        dict: items 'error' and 'associated_key'
 
-        # None: all keys were present
-        # dict: items 'error' and 'associated_key'
-
-        # Assume all keys are present.
+        Assume all keys are present.
+        """
         missing_keys = []
 
         for current_key in key_set:
